Bring plugin args from _internal to the user facing Plugin class
